Moves item children helper into item children list folder
