+++ src/app/modules/item/components/chapter-children/chapter-children.component.ts
@@ -10,7 +10,7 @@ import { mapToFetchState } from '../../../../shared/operators/state';
 import { canCurrentUserViewContent } from 'src/app/shared/models/domain/item-view-permission';
 import { GroupWatchingService } from 'src/app/core/services/group-watching.service';
 import { LayoutService } from '../../../../shared/services/layout.service';
-import { ItemChildWithAdditions } from '../../helpers/item-children';
+import { ItemChildWithAdditions } from '../item-children-list/item-children';
 
 @Component({
   selector: 'alg-chapter-children',
